+++ mobile/src/main/java/com/aptopayments/mobile/repository/user/remote/UserApi.kt
@@ -4,6 +4,7 @@ import com.aptopayments.mobile.network.X_AUTHORIZATION
 import com.aptopayments.mobile.repository.user.remote.entities.NotificationPreferencesEntity
 import com.aptopayments.mobile.repository.user.remote.entities.UserEntity
 import com.aptopayments.mobile.repository.user.remote.requests.*
+import okhttp3.RequestBody
 import retrofit2.Call
 import retrofit2.http.*
 
@@ -20,7 +21,11 @@ private const val PUSH_TOKEN = "push_token"
 internal interface UserApi {
 
     @POST(CREATE_USER_PATH)
-    fun createUser(@Body request: CreateUserDataRequest): Call<UserEntity>
+    fun createUserSigned(@Body request: CreateUserDataRequest): Call<UserEntity>
+
+    @POST(CREATE_USER_PATH)
+    @Headers("Content-Type: application/jwt")
+    fun createUser(@Body webToken: RequestBody? = null): Call<UserEntity>
 
     @PUT(USER_PATH)
     fun updateUser(@Body request: UserDataRequest): Call<UserEntity>